+++ src/main/Exception/InvalidArgumentException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace WebArch\BitrixUserPropertyType\Exception;
+
+use InvalidArgumentException as CommonInvalidArgumentException;
+
+class InvalidArgumentException extends CommonInvalidArgumentException implements BitrixUserTypeExceptionInterface
+{
+}