+++ src/main/Exception/LogicException.php
@@ -4,6 +4,6 @@ namespace WebArch\BitrixUserPropertyType\Exception;
 
 use LogicException as CommonLogicException;
 
-class LogicException extends CommonLogicException
+class LogicException extends CommonLogicException implements BitrixUserTypeExceptionInterface
 {
 }